= 1.2.0 = * Added debugging options in plugin admin. * Log to Wordpress debug.log - Requires WP_DEBUG & WP_DEBUG_LOG set to true in wp-config.php * Added user friendly error messages for public facing screens. *  Generic apology / error message. *  Add specific error to generic message - Set WP_DEBUG_DISPLAY to false in wp-config.php

+++ includes/class-eventz-activator.php
@@ -22,8 +22,9 @@ class Eventz_Lite_Activator {
         $default_options = array('_version'=>$version,'_username'=>'','_password'=>'',
             '_endpoint'=>'api.eventfinda.co.nz','_event_location'=>'1','_event_date'=>'1',
                 '_event_category'=>'1','_event_excerpt'=>'300','_event_separator'=>'1',
-                    '_results_pp'=>'10','_delete_options'=>'0','_eventfinda_logo'=>'1',
-                        '_eventfinda_text'=>'0','_show_plugin_logo'=>'0','_show_plugin_link'=>'0');
+                    '_results_pp'=>'10','_debug'=>'0','_debug_screen'=>'0','_delete_options'=>'0',
+                        '_eventfinda_logo'=>'1','_eventfinda_text'=>'0','_show_plugin_logo'=>'0',
+                            '_show_plugin_link'=>'0');
         $options = get_option(self::$option_name, array());
         if(!$options) {
             update_option(self::$option_name, $default_options);
@@ -42,6 +43,14 @@ class Eventz_Lite_Activator {
                     '_event_category'=>'1',
                     '_event_excerpt'=>'300',
                     '_event_separator'=>'1'
+                );
+                update_option(self::$option_name, array_merge($new_options, $options));
+            }
+            $new_key = '_debug';
+            if (!array_key_exists($new_key, $options)) {
+                /* Probaly safe to assume other keys do not exist so update options */
+                $new_options = array('_debug'=>'0',
+                    '_debug_screen'=>'0'
                 );
                 update_option(self::$option_name, array_merge($new_options, $options));
             }
